# Conflicts:
#	src/Nethermind/Nethermind.Blockchain.Test/Validators/BlockValidatorTests.cs
#	src/Nethermind/Nethermind.JsonRpc.Test/Modules/Eth/EthRpcModuleTests.cs

+++ src/Nethermind/Nethermind.Trie/TrieNode.cs
@@ -84,7 +84,7 @@ namespace Nethermind.Trie
             while (true)
             {
                 ulong current = Volatile.Read(ref _rlpSeqAndLength);
-                uint seq = (uint)(current >> 32);
+                ulong seq = current >> 32;
                 if ((seq & 1) != 0)
                 {
                     // Another writer is active — spin until it completes
@@ -92,13 +92,13 @@ namespace Nethermind.Trie
                     continue;
                 }
                 // Set lock bit (odd) — seq | 1 is always odd regardless of overflow
-                ulong writing = (ulong)(seq | 1) << 32;
+                ulong writing = (seq | 1) << 32;
                 if (Interlocked.CompareExchange(ref _rlpSeqAndLength, writing, current) == current)
                 {
                     Volatile.Write(ref _rlpArray, value.UnderlyingArray);
                     // Advance sequence by 2 and clear lock bit (even), store final length
-                    uint doneSeq = (seq + 2) & 0xFFFFFFFE;
-                    Volatile.Write(ref _rlpSeqAndLength, (ulong)doneSeq << 32 | (uint)value.Length);
+                    ulong doneSeq = (seq + 2) & 0xFFFFFFFE;
+                    Volatile.Write(ref _rlpSeqAndLength, doneSeq << 32 | (uint)value.Length);
                     return;
                 }
                 spin.SpinOnce(); // CAS failed — another writer raced; back off before retry
